requirements.txt
faiss-cpu>=1.13.2
gensim>=4.4.0
ipykernel>=7.1.0
nltk>=3.9.2
pymorphy3>=2.0.6
scikit-learn>=1.8.0
sentence-transformers>=5.2.0
setuptools>=80.9.0
tqdm>=4.67.1
yargy>=0.16.0